refactor: Separate Post contents component
