+++ src/components/Post.css
@@ -51,18 +51,6 @@
   cursor: pointer;
 }
 
-.post__contents {
-  font-size: var(--font-size-xs);
-  margin-bottom: var(--margin-s);
-
-  white-space: pre-line;
-}
-
-.post__contents--more {
-  margin-top: var(--margin-xs);
-  cursor: pointer;
-}
-
 @media screen and (max-width: 900px) {
   .post {
     flex-direction: column;
